🎉 feat(Seo.astro): add SEO component and siteData.json The SEO component has been added to the project to improve search engine optimization. The component generates meta tags for Open Graph and Twitter, as well as a JSON-LD script for structured data. The siteData.json file has also been added to store site-wide data such as the site title, description, and default image.
🆕 feat(jsonLD.js): add support for generating JSON-LD for website
🐛 fix(BlogPostLayout.astro): add missing frontmatter props to MainLayout component
The jsonLDGenerator function now supports generating JSON-LD for a website in addition to blog posts. The website JSON-LD includes the website name and URL. The BlogPostLayout.astro file was missing some frontmatter props in the MainLayout component, which caused issues with generating JSON-LD for blog posts.

🎨 style(MainHead.astro): add Seo component to MainHead
🎨 style(MainLayout.astro): update title, description, and add image, frontmatter, and robots props
🎨 style(index.astro): increase headshot width to 330px
The Seo component is added to MainHead to improve SEO by providing metadata for search engines. The title and description are updated in MainLayout to better reflect the purpose of the website. The image, frontmatter, and robots props are added to provide additional metadata for search engines. The headshot width is increased to 330px to improve the visual appearance of the website.