🎨 style(utils.js): change date format to use Belgian locale and Central European Timezone 🗑️ chore(utils.js): remove unused slugify function 🖼️ feat(portfolio.jpg): add new portfolio image The formatDate function now uses the Belgian locale and Central European Timezone, which is more appropriate for the application's target audience. The slugify function was removed as it was no longer used in the application. A new portfolio image was added to the public/images directory.
🎨 style(blog): update post-6.md with project description and why Astro was chosen
This commit updates the post-6.md file with a project description and why Astro was chosen as the technology to build the portfolio. The project is a personal portfolio website built with Astro, which allows for easy updates and maintenance. Astro was chosen because it automatically generates routes, works with components, allows for the use of common frameworks and libraries, and enables dynamic pagination.

🎨 style(category.astro): update category names
The category names have been updated to better reflect the content of the pages. The changes include updating the category names for Frontend, Backend, and Laravel.