🎨 style(BlogPostLayout.astro): add a link to the GitHub repository in the blog post layout 🗑️ chore(image-default.jpg): delete old default image and add new default image The blog post layout now includes a link to the GitHub repository associated with the blog post. This provides readers with an easy way to access the code associated with the post. The old default image was deleted and replaced with a new default image.
📝 docs(post-1.md): add Github repository link
Added a link to the Github repository for the Laravel and Beyond project. This will make it easier for others to access the code and contribute to the project.

📝 docs(blog): add Github links to blog posts
Added Github links to two blog posts. The first post now has a Github link to the repository where the project is hosted. The second post now has a Github link to the repository where the author's other projects are hosted.

🚀 feat(pages): add Full Stack category and update title of index page
The Full Stack category has been added to the category page with the name "Full Stack". The index page title has been updated to "Thomas Cano Morant Showcase" to better reflect the purpose of the website.