🎨 style(about.astro): reformat code and fix indentation The code has been reformatted to improve readability and consistency. Indentation has been fixed to ensure that the code is properly aligned.
🚀 chore(about.astro, index.astro): update content
The content of the about page has been updated to include a link to download my resume. The content of the index page has been updated to use h3 tags instead of p tags for better semantic structure.

🎨 style(global.css): add italic class
This commit adds a new class `.italic` to the global.css file. The `.italic` class sets the font-style property to italic.