🎉 feat(intro_react.jpg): add introductory image for React An introductory image for React has been added to the public/images directory. This image will be used to provide a visual representation of React in the application.
🚀 chore(post-7.md): update blog post with GF Todo App project
📝 docs(post-7.md): add details about the GF Todo App project
This commit updates the blog post to reflect the creation of the GF Todo App project. The post now includes details about the project, including the use of React components, the creation of a TodoList component, and the creation of a Table component. The post also includes code examples and explanations of how the components work.

✨ feat(category.astro): add WIP category
A new category has been added to the list of categories. The category is called WIP and has been added to the getStaticPaths function.