🎉 feat(post-8.jpg): add new image A new image has been added to the public/images directory. The image is named post-8.jpg and is now available for use in the application.
🎉 feat(blog): add blog post 8 - Puntopia
This commit adds a new blog post, "Puntopia", which describes an open-source community sharing tool designed to discover, create, and share funny puns. The post includes information on the project's requirements, technologies used, file structure, and more. It also provides a detailed breakdown of the Next.js routing system and the Pun and User models' schemas. Finally, it includes an `.env.example` file to provide the basic structure for the environment file.